pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[project]
name = "mysql_editor_python"
version = "1.0"
description = "A MySQL Editor"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pyside6>=6.6.0",
    "mysql-connector-python>=8.0.31",
    "requests>=2.31.0"
]
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: Microsoft :: Windows",
    "Operating System :: POSIX :: Linux",
]

[project.gui-scripts]
mysql-editor = "mysql_editor:main"

[project.urls]
Homepage = "https://github.com/PandaRules/MySQL-Editor-Python"
Issues = "https://github.com/PandaRules/MySQL-Editor-Python/issues"
